refactoring boundary element layout
