+++ engeom/src/geom2/boundary.rs
@@ -44,33 +44,3 @@ impl BoundaryElement for Segment2 {
         SurfacePoint2::new_normalize(self.at(fraction), self.orthogonal())
     }
 }
-
-impl BoundaryElement for Arc2 {
-    fn closest_to_point(&self, point: &impl PCoords<2>) -> f64 {
-        let theta = self.circle.angle_of_point(point);
-        if self.is_theta_on_arc(theta) {
-            self.theta_to_fraction(theta) * self.length()
-        } else {
-            let d0 = dist(&self.at_start(), point);
-            let d1 = dist(&self.at_end(), point);
-            if d0 < d1 { 0.0 } else { self.length() }
-        }
-    }
-
-    fn aabb(&self) -> Aabb2 {
-        Arc2::aabb(self)
-    }
-
-    fn at_start(&self) -> SurfacePoint2 {
-        self.at_length(0.0)
-    }
-
-    fn at_end(&self) -> SurfacePoint2 {
-        self.at_length(self.length())
-    }
-
-    fn at_length(&self, length: f64) -> SurfacePoint2 {
-        let p = self.point_at_length(length);
-        SurfacePoint2::new_normalize(p, p - self.center())
-    }
-}
